+++ basic/controllers/SubmittedFlightPlanController.php
@@ -479,7 +479,8 @@ class SubmittedFlightPlanController extends Controller
                     $model->cargo_paid_kg,
                     CK::getPaxAdultWeightKg(),
                     CK::getPaxChildWeightKg(),
-                    CK::getPaxCheckedBaggageKg()
+                    CK::getPaxCheckedBaggageKg(),
+                    $model->aircraft?->aircraftConfiguration?->oew
                 );
             }